Remove redundant raw markdown files for models with generated docs
