--- docs/raw/connectors/abuseipdb.md
@@ -1,12 +1,0 @@
-
-AbuseIPDB
-=========
-
-
-
-api_key
--------
-
-- Type: `secret` (required)
-
-The unique AbuseIPDB API key.
