--- docs/raw/connectors/amplitude.md
@@ -1,32 +1,0 @@
-
-Amplitude
-=========
-
-
-
-api_key
--------
-
-- Type: `secret` (required)
-
-The Amplitude API Key generated for the Descope service.
-
-
-
-server_url
-----------
-
-- Type: `string` 
-
-The server URL of the Amplitude API, when using different api or a custom domain
-in Amplitude.
-
-
-
-server_zone
------------
-
-- Type: `string` 
-
-`EU` or `US`. Sets the Amplitude server zone. Set this to `EU` for Amplitude
-projects created in `EU` data center. Default is `US`.
