--- docs/raw/connectors/awstranslate.md
@@ -1,40 +1,0 @@
-
-AWSTranslate
-============
-
-
-
-access_key_id
--------------
-
-- Type: `string` (required)
-
-AWS access key ID.
-
-
-
-secret_access_key
------------------
-
-- Type: `secret` (required)
-
-AWS secret access key.
-
-
-
-session_token
--------------
-
-- Type: `secret` 
-
-(Optional) A security or session token to use with these credentials. Usually
-present for temporary credentials.
-
-
-
-region
-------
-
-- Type: `string` (required)
-
-The AWS region to which this client will send requests. (e.g. us-east-1.)
