--- docs/raw/connectors/clear.md
@@ -1,21 +1,0 @@
-
-Clear
-=====
-
-
-
-project_id
-----------
-
-- Type: `string` (required)
-
-Your CLEAR project ID.
-
-
-
-api_key
--------
-
-- Type: `secret` (required)
-
-Your CLEAR API key.
