--- docs/raw/connectors/datadog.md
@@ -1,50 +1,0 @@
-
-Datadog
-=======
-
-
-
-api_key
--------
-
-- Type: `secret` (required)
-
-The unique Datadog organization key.
-
-
-
-site
-----
-
-- Type: `string` 
-
-The Datadog site to send logs to. Default is `datadoghq.com`. European, free
-tier and other customers should set their site accordingly.
-
-
-
-audit_enabled
--------------
-
-- Type: `bool` 
-- Default: `true`
-
-// description for audit_enabled
-
-
-
-audit_filters
--------------
-
-- Type: `string` 
-
-// description for audit_filters
-
-
-
-troubleshoot_log_enabled
-------------------------
-
-- Type: `bool` 
-
-// description for troubleshoot_log_enabled
