--- docs/raw/connectors/devrevgrow.md
@@ -1,12 +1,0 @@
-
-DevRevGrow
-==========
-
-
-
-api_key
--------
-
-- Type: `secret` (required)
-
-Authentication to DevRev APIs requires a personal access token (PAT).
