--- docs/raw/connectors/docebo.md
@@ -1,48 +1,0 @@
-
-Docebo
-======
-
-
-
-base_url
---------
-
-- Type: `string` (required)
-
-The Docebo api base url.
-
-
-
-client_id
----------
-
-- Type: `string` (required)
-
-The Docebo OAuth 2.0 app client ID.
-
-
-
-client_secret
--------------
-
-- Type: `secret` (required)
-
-The Docebo OAuth 2.0 app client secret.
-
-
-
-username
---------
-
-- Type: `string` (required)
-
-The Docebo username.
-
-
-
-password
---------
-
-- Type: `secret` (required)
-
-The Docebo user's password.
