--- docs/raw/connectors/fingerprint.md
@@ -1,49 +1,0 @@
-
-Fingerprint
-===========
-
-
-
-public_api_key
---------------
-
-- Type: `string` (required)
-
-The Fingerprint public API key.
-
-
-
-secret_api_key
---------------
-
-- Type: `secret` (required)
-
-The Fingerprint secret API key.
-
-
-
-use_cloudflare_integration
---------------------------
-
-- Type: `bool` 
-
-Enable to configure the relevant Cloudflare integration parameters if Cloudflare
-integration is set in your Fingerprint account.
-
-
-
-cloudflare_script_url
----------------------
-
-- Type: `string` 
-
-The Cloudflare integration Script URL.
-
-
-
-cloudflare_endpoint_url
------------------------
-
-- Type: `string` 
-
-The Cloudflare integration Endpoint URL.
