--- docs/raw/connectors/fingerprintdescope.md
@@ -1,12 +1,0 @@
-
-FingerprintDescope
-==================
-
-
-
-custom_domain
--------------
-
-- Type: `string` 
-
-The custom domain to fetch
