--- docs/raw/connectors/forter.md
@@ -1,51 +1,0 @@
-
-Forter
-======
-
-
-
-site_id
--------
-
-- Type: `string` (required)
-
-The Forter site ID.
-
-
-
-secret_key
-----------
-
-- Type: `secret` (required)
-
-The Forter secret key.
-
-
-
-overrides
----------
-
-- Type: `bool` 
-
-Override the user's IP address or email so that Forter can provide a specific
-decision or recommendation. Contact the Forter team for further details. Note:
-Overriding the user IP address or email is intended for testing purpose and
-should not be utilized in production environments.
-
-
-
-override_ip_address
--------------------
-
-- Type: `string` 
-
-Override the user IP address.
-
-
-
-override_user_email
--------------------
-
-- Type: `string` 
-
-Override the user email.
