--- docs/raw/connectors/googlecloudtranslation.md
@@ -1,21 +1,0 @@
-
-GoogleCloudTranslation
-======================
-
-
-
-project_id
-----------
-
-- Type: `string` (required)
-
-The Google Cloud project ID where the Google Cloud Translation is managed.
-
-
-
-service_account_json
---------------------
-
-- Type: `secret` (required)
-
-Service Account JSON associated with the current project.
