--- docs/raw/connectors/hibp.md
@@ -1,3 +1,0 @@
-
-HIBP
-====
