--- docs/raw/connectors/hubspot.md
@@ -1,22 +1,0 @@
-
-HubSpot
-=======
-
-
-
-access_token
-------------
-
-- Type: `secret` (required)
-
-The HubSpot private API access token generated for the Descope service.
-
-
-
-base_url
---------
-
-- Type: `string` 
-
-The base URL of the HubSpot API, when using a custom domain in HubSpot, default
-value is https://api.hubapi.com .
