--- docs/raw/connectors/intercom.md
@@ -1,22 +1,0 @@
-
-Intercom
-========
-
-
-
-token
------
-
-- Type: `secret` (required)
-
-The Intercom access token.
-
-
-
-region
-------
-
-- Type: `string` 
-- Default: `"US"`
-
-Regional Hosting - US, EU, or AU. default: US
