--- docs/raw/connectors/recaptcha.md
@@ -1,23 +1,0 @@
-
-Recaptcha
-=========
-
-
-
-site_key
---------
-
-- Type: `string` (required)
-
-The site key is used to invoke reCAPTCHA service on your site or mobile
-application.
-
-
-
-secret_key
-----------
-
-- Type: `secret` (required)
-
-The secret key authorizes communication between Descope backend and the
-reCAPTCHA server to verify the user's response.
