--- docs/raw/connectors/rekognition.md
@@ -1,31 +1,0 @@
-
-Rekognition
-===========
-
-
-
-access_key_id
--------------
-
-- Type: `string` (required)
-
-The AWS access key ID
-
-
-
-secret_access_key
------------------
-
-- Type: `secret` (required)
-
-The AWS secret access key
-
-
-
-collection_id
--------------
-
-- Type: `string` (required)
-
-The collection to store registered users in. Should match `[a-zA-Z0-9_.-]+`
-pattern. Changing this will cause losing existing users.
