--- docs/raw/connectors/salesforce.md
@@ -1,39 +1,0 @@
-
-Salesforce
-==========
-
-
-
-base_url
---------
-
-- Type: `string` (required)
-
-The Salesforce API base URL.
-
-
-
-client_id
----------
-
-- Type: `string` (required)
-
-The consumer key of the connected app.
-
-
-
-client_secret
--------------
-
-- Type: `secret` (required)
-
-The consumer secret of the connected app.
-
-
-
-version
--------
-
-- Type: `string` (required)
-
-REST API Version.
