--- docs/raw/connectors/segment.md
@@ -1,21 +1,0 @@
-
-Segment
-=======
-
-
-
-write_key
----------
-
-- Type: `secret` (required)
-
-The Segment Write Key generated for the Descope service.
-
-
-
-host
-----
-
-- Type: `string` 
-
-The base URL of the Segment API, when using a custom domain in Segment.
