--- docs/raw/connectors/sumologic.md
@@ -1,40 +1,0 @@
-
-SumoLogic
-=========
-
-
-
-http_source_url
----------------
-
-- Type: `secret` (required)
-
-The URL associated with an HTTP Hosted collector
-
-
-
-audit_enabled
--------------
-
-- Type: `bool` 
-- Default: `true`
-
-// description for audit_enabled
-
-
-
-audit_filters
--------------
-
-- Type: `string` 
-
-// description for audit_filters
-
-
-
-troubleshoot_log_enabled
-------------------------
-
-- Type: `bool` 
-
-// description for troubleshoot_log_enabled
