--- docs/raw/connectors/telesign.md
@@ -1,21 +1,0 @@
-
-Telesign
-========
-
-
-
-customer_id
------------
-
-- Type: `string` (required)
-
-The unique Telesign account Customer ID
-
-
-
-api_key
--------
-
-- Type: `secret` (required)
-
-The unique Telesign API key
