--- docs/raw/connectors/traceable.md
@@ -1,21 +1,0 @@
-
-Traceable
-=========
-
-
-
-secret_key
-----------
-
-- Type: `secret` (required)
-
-The Traceable secret key.
-
-
-
-eu_region
----------
-
-- Type: `bool` 
-
-EU(Europe) Region deployment of Traceable platform.
