--- docs/raw/connectors/veriff.md
@@ -1,30 +1,0 @@
-
-Veriff
-======
-
-
-
-api_key
--------
-
-- Type: `string` (required)
-
-The Veriff Public API Key, you can find under Veriff Station - Integrations.
-
-
-
-secret_key
-----------
-
-- Type: `secret` (required)
-
-The Veriff Shared secret key, you can find under Veriff Station - Integrations.
-
-
-
-base_url
---------
-
-- Type: `string` 
-
-The base URL of the Veriff API, default value is https://stationapi.veriff.com.
