+++ tools/terragen/schema/model.go
@@ -1,0 +1,8 @@
+package schema
+
+type Model struct {
+	Name      string
+	Package   string
+	Fields    []*Field
+	Generated bool
+}
